Reorganizing and improving the rendering system.
Moving logical game object renderers out of the game-env and into a
GameRenderer record type. The record itself implements the PRender
protocol which does proper matrix transformations for locating objects
in the rendering context. This also allows for the removal of certain
extraneous values from the game object renderers, e.g. the border size
used when rendering the player.