Changing exit representation so it's embedded in the board.
Previously, the game had a separate exit object that was essentially
overlaid on the map. Now the board maintains an exit-room field. The
board renderer also changes to carry a vector of room renderers; one
for each room object, kept in parallel. The board renderer takes the
board as an argument and creates the room renderers appropriately,
currently creating one with a different color for the exit room.